requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-cov==4.1.0
allure-pytest==2.13.2
requests==2.31.0
jsonschema==4.20.0
Faker==20.1.0
python-dotenv==1.0.0
colorlog==6.8.0
